python/pyproject.toml
[tool.poetry]
name = "magneticTweezers"
version = "0.1.0"
description = ""
authors = ["Erik Plesko <erik.plesko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
numpy = "^1.19"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"
matplotlib = "^3.5.2"
jupyterthemes = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
